+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 
 - Uniformize keyboard scancode values accross Wayland and X11 (#297).
+- Internal rework of the wayland event loop
+- Added method `os::linux::WindowExt::is_ready`
 
 # Version 0.8.1 (2017-09-22)